requirements.txt
crewai
langchain_openai
python-dotenv
python-docx
pydantic
crewai_tools
scikit-learn
numpy
joblib
docx2python
sentence-transformers